new file structure -- how do we feel?
