+++ server/app.js
@@ -1,0 +1,15 @@
+require('dotenv').config();
+const express = require('express');
+const bodyParser = require('body-parser');
+const routes = require('./routes/routes');
+const app = express();
+
+app.use(bodyParser.json());
+routes(app);
+
+app.use((err, req, res, next) => {
+  console.log(err);
+  res.status(422).send({ error: err.message });
+});
+
+module.exports = app;
